requirements.txt
pytest
Appium-Python-Client
selenium
opencv-python
numpy